requirements.txt
certifi==2020.11.8
chardet==3.0.4
idna==2.10
oauthlib==3.1.0
pkg-resources==0.0.0
PySocks==1.7.1
python-dotenv==0.15.0
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
tweepy==3.9.0
urllib3==1.26.2
